sdk/typescript: unify all models into a single file
