--- sdk/generator/typescript/template/src/version/models/index.ts
@@ -1,3 +1,0 @@
-export * from "./inputs";
-export * from "./outputs";
-export * from "./literals";
