--- sdk/typescript/src/2026-04/models/index.ts
@@ -1,3 +1,0 @@
-export * from "./inputs";
-export * from "./outputs";
-export * from "./literals";
