--- sdk/typescript/src/2026-04/models/literals.ts
@@ -1,1909 +1,0 @@
-/**
- * AggregationFunction
- */
-export type AggregationFunction = "count" | "sum" | "max" | "min" | "avg" | "unique";
-/**
- * BenefitGrantSortProperty
- */
-export type BenefitGrantSortProperty =
-  | "created_at"
-  | "-created_at"
-  | "granted_at"
-  | "-granted_at"
-  | "revoked_at"
-  | "-revoked_at";
-/**
- * BenefitSortProperty
- */
-export type BenefitSortProperty =
-  | "created_at"
-  | "-created_at"
-  | "description"
-  | "-description"
-  | "type"
-  | "-type"
-  | "user_order"
-  | "-user_order";
-/**
- * BenefitType
- */
-export type BenefitType =
-  | "custom"
-  | "discord"
-  | "github_repository"
-  | "downloadables"
-  | "license_keys"
-  | "meter_credit"
-  | "feature_flag"
-  | "slack_shared_channel";
-/**
- * BenefitVisibility
- */
-export type BenefitVisibility = "draft" | "private" | "public";
-/**
- * BillingAddressFieldMode
- */
-export type BillingAddressFieldMode = "required" | "optional" | "disabled";
-/**
- * CheckoutLinkSortProperty
- */
-export type CheckoutLinkSortProperty =
-  | "created_at"
-  | "-created_at"
-  | "label"
-  | "-label"
-  | "success_url"
-  | "-success_url"
-  | "allow_discount_codes"
-  | "-allow_discount_codes";
-/**
- * CheckoutSortProperty
- */
-export type CheckoutSortProperty =
-  | "created_at"
-  | "-created_at"
-  | "expires_at"
-  | "-expires_at"
-  | "status"
-  | "-status";
-/**
- * CheckoutStatus
- */
-export type CheckoutStatus = "open" | "expired" | "confirmed" | "succeeded" | "failed";
-/**
- * CountryAlpha2
- */
-export type CountryAlpha2 =
-  | "AD"
-  | "AE"
-  | "AF"
-  | "AG"
-  | "AI"
-  | "AL"
-  | "AM"
-  | "AO"
-  | "AQ"
-  | "AR"
-  | "AS"
-  | "AT"
-  | "AU"
-  | "AW"
-  | "AX"
-  | "AZ"
-  | "BA"
-  | "BB"
-  | "BD"
-  | "BE"
-  | "BF"
-  | "BG"
-  | "BH"
-  | "BI"
-  | "BJ"
-  | "BL"
-  | "BM"
-  | "BN"
-  | "BO"
-  | "BQ"
-  | "BR"
-  | "BS"
-  | "BT"
-  | "BV"
-  | "BW"
-  | "BY"
-  | "BZ"
-  | "CA"
-  | "CC"
-  | "CD"
-  | "CF"
-  | "CG"
-  | "CH"
-  | "CI"
-  | "CK"
-  | "CL"
-  | "CM"
-  | "CN"
-  | "CO"
-  | "CR"
-  | "CU"
-  | "CV"
-  | "CW"
-  | "CX"
-  | "CY"
-  | "CZ"
-  | "DE"
-  | "DJ"
-  | "DK"
-  | "DM"
-  | "DO"
-  | "DZ"
-  | "EC"
-  | "EE"
-  | "EG"
-  | "EH"
-  | "ER"
-  | "ES"
-  | "ET"
-  | "FI"
-  | "FJ"
-  | "FK"
-  | "FM"
-  | "FO"
-  | "FR"
-  | "GA"
-  | "GB"
-  | "GD"
-  | "GE"
-  | "GF"
-  | "GG"
-  | "GH"
-  | "GI"
-  | "GL"
-  | "GM"
-  | "GN"
-  | "GP"
-  | "GQ"
-  | "GR"
-  | "GS"
-  | "GT"
-  | "GU"
-  | "GW"
-  | "GY"
-  | "HK"
-  | "HM"
-  | "HN"
-  | "HR"
-  | "HT"
-  | "HU"
-  | "ID"
-  | "IE"
-  | "IL"
-  | "IM"
-  | "IN"
-  | "IO"
-  | "IQ"
-  | "IR"
-  | "IS"
-  | "IT"
-  | "JE"
-  | "JM"
-  | "JO"
-  | "JP"
-  | "KE"
-  | "KG"
-  | "KH"
-  | "KI"
-  | "KM"
-  | "KN"
-  | "KP"
-  | "KR"
-  | "KW"
-  | "KY"
-  | "KZ"
-  | "LA"
-  | "LB"
-  | "LC"
-  | "LI"
-  | "LK"
-  | "LR"
-  | "LS"
-  | "LT"
-  | "LU"
-  | "LV"
-  | "LY"
-  | "MA"
-  | "MC"
-  | "MD"
-  | "ME"
-  | "MF"
-  | "MG"
-  | "MH"
-  | "MK"
-  | "ML"
-  | "MM"
-  | "MN"
-  | "MO"
-  | "MP"
-  | "MQ"
-  | "MR"
-  | "MS"
-  | "MT"
-  | "MU"
-  | "MV"
-  | "MW"
-  | "MX"
-  | "MY"
-  | "MZ"
-  | "NA"
-  | "NC"
-  | "NE"
-  | "NF"
-  | "NG"
-  | "NI"
-  | "NL"
-  | "NO"
-  | "NP"
-  | "NR"
-  | "NU"
-  | "NZ"
-  | "OM"
-  | "PA"
-  | "PE"
-  | "PF"
-  | "PG"
-  | "PH"
-  | "PK"
-  | "PL"
-  | "PM"
-  | "PN"
-  | "PR"
-  | "PS"
-  | "PT"
-  | "PW"
-  | "PY"
-  | "QA"
-  | "RE"
-  | "RO"
-  | "RS"
-  | "RU"
-  | "RW"
-  | "SA"
-  | "SB"
-  | "SC"
-  | "SD"
-  | "SE"
-  | "SG"
-  | "SH"
-  | "SI"
-  | "SJ"
-  | "SK"
-  | "SL"
-  | "SM"
-  | "SN"
-  | "SO"
-  | "SR"
-  | "SS"
-  | "ST"
-  | "SV"
-  | "SX"
-  | "SY"
-  | "SZ"
-  | "TC"
-  | "TD"
-  | "TF"
-  | "TG"
-  | "TH"
-  | "TJ"
-  | "TK"
-  | "TL"
-  | "TM"
-  | "TN"
-  | "TO"
-  | "TR"
-  | "TT"
-  | "TV"
-  | "TW"
-  | "TZ"
-  | "UA"
-  | "UG"
-  | "UM"
-  | "US"
-  | "UY"
-  | "UZ"
-  | "VA"
-  | "VC"
-  | "VE"
-  | "VG"
-  | "VI"
-  | "VN"
-  | "VU"
-  | "WF"
-  | "WS"
-  | "YE"
-  | "YT"
-  | "ZA"
-  | "ZM"
-  | "ZW";
-/**
- * CountryAlpha2Input
- */
-export type CountryAlpha2Input =
-  | "AD"
-  | "AE"
-  | "AF"
-  | "AG"
-  | "AI"
-  | "AL"
-  | "AM"
-  | "AO"
-  | "AQ"
-  | "AR"
-  | "AS"
-  | "AT"
-  | "AU"
-  | "AW"
-  | "AX"
-  | "AZ"
-  | "BA"
-  | "BB"
-  | "BD"
-  | "BE"
-  | "BF"
-  | "BG"
-  | "BH"
-  | "BI"
-  | "BJ"
-  | "BL"
-  | "BM"
-  | "BN"
-  | "BO"
-  | "BQ"
-  | "BR"
-  | "BS"
-  | "BT"
-  | "BV"
-  | "BW"
-  | "BY"
-  | "BZ"
-  | "CA"
-  | "CC"
-  | "CD"
-  | "CF"
-  | "CG"
-  | "CH"
-  | "CI"
-  | "CK"
-  | "CL"
-  | "CM"
-  | "CN"
-  | "CO"
-  | "CR"
-  | "CV"
-  | "CW"
-  | "CX"
-  | "CY"
-  | "CZ"
-  | "DE"
-  | "DJ"
-  | "DK"
-  | "DM"
-  | "DO"
-  | "DZ"
-  | "EC"
-  | "EE"
-  | "EG"
-  | "EH"
-  | "ER"
-  | "ES"
-  | "ET"
-  | "FI"
-  | "FJ"
-  | "FK"
-  | "FM"
-  | "FO"
-  | "FR"
-  | "GA"
-  | "GB"
-  | "GD"
-  | "GE"
-  | "GF"
-  | "GG"
-  | "GH"
-  | "GI"
-  | "GL"
-  | "GM"
-  | "GN"
-  | "GP"
-  | "GQ"
-  | "GR"
-  | "GS"
-  | "GT"
-  | "GU"
-  | "GW"
-  | "GY"
-  | "HK"
-  | "HM"
-  | "HN"
-  | "HR"
-  | "HT"
-  | "HU"
-  | "ID"
-  | "IE"
-  | "IL"
-  | "IM"
-  | "IN"
-  | "IO"
-  | "IQ"
-  | "IS"
-  | "IT"
-  | "JE"
-  | "JM"
-  | "JO"
-  | "JP"
-  | "KE"
-  | "KG"
-  | "KH"
-  | "KI"
-  | "KM"
-  | "KN"
-  | "KR"
-  | "KW"
-  | "KY"
-  | "KZ"
-  | "LA"
-  | "LB"
-  | "LC"
-  | "LI"
-  | "LK"
-  | "LR"
-  | "LS"
-  | "LT"
-  | "LU"
-  | "LV"
-  | "LY"
-  | "MA"
-  | "MC"
-  | "MD"
-  | "ME"
-  | "MF"
-  | "MG"
-  | "MH"
-  | "MK"
-  | "ML"
-  | "MM"
-  | "MN"
-  | "MO"
-  | "MP"
-  | "MQ"
-  | "MR"
-  | "MS"
-  | "MT"
-  | "MU"
-  | "MV"
-  | "MW"
-  | "MX"
-  | "MY"
-  | "MZ"
-  | "NA"
-  | "NC"
-  | "NE"
-  | "NF"
-  | "NG"
-  | "NI"
-  | "NL"
-  | "NO"
-  | "NP"
-  | "NR"
-  | "NU"
-  | "NZ"
-  | "OM"
-  | "PA"
-  | "PE"
-  | "PF"
-  | "PG"
-  | "PH"
-  | "PK"
-  | "PL"
-  | "PM"
-  | "PN"
-  | "PR"
-  | "PS"
-  | "PT"
-  | "PW"
-  | "PY"
-  | "QA"
-  | "RE"
-  | "RO"
-  | "RS"
-  | "RW"
-  | "SA"
-  | "SB"
-  | "SC"
-  | "SD"
-  | "SE"
-  | "SG"
-  | "SH"
-  | "SI"
-  | "SJ"
-  | "SK"
-  | "SL"
-  | "SM"
-  | "SN"
-  | "SO"
-  | "SR"
-  | "SS"
-  | "ST"
-  | "SV"
-  | "SX"
-  | "SZ"
-  | "TC"
-  | "TD"
-  | "TF"
-  | "TG"
-  | "TH"
-  | "TJ"
-  | "TK"
-  | "TL"
-  | "TM"
-  | "TN"
-  | "TO"
-  | "TR"
-  | "TT"
-  | "TV"
-  | "TW"
-  | "TZ"
-  | "UA"
-  | "UG"
-  | "UM"
-  | "US"
-  | "UY"
-  | "UZ"
-  | "VA"
-  | "VC"
-  | "VE"
-  | "VG"
-  | "VI"
-  | "VN"
-  | "VU"
-  | "WF"
-  | "WS"
-  | "YE"
-  | "YT"
-  | "ZA"
-  | "ZM"
-  | "ZW";
-/**
- * CustomFieldSortProperty
- */
-export type CustomFieldSortProperty =
-  | "created_at"
-  | "-created_at"
-  | "slug"
-  | "-slug"
-  | "name"
-  | "-name"
-  | "type"
-  | "-type";
-/**
- * CustomFieldType
- */
-export type CustomFieldType = "text" | "number" | "date" | "checkbox" | "select";
-/**
- * CustomerBenefitGrantSortProperty
- */
-export type CustomerBenefitGrantSortProperty =
-  | "granted_at"
-  | "-granted_at"
-  | "type"
-  | "-type"
-  | "organization"
-  | "-organization"
-  | "product_benefit"
-  | "-product_benefit";
-/**
- * CustomerCancellationReason
- */
-export type CustomerCancellationReason =
-  | "customer_service"
-  | "low_quality"
-  | "missing_features"
-  | "switched_service"
-  | "too_complex"
-  | "too_expensive"
-  | "unused"
-  | "other";
-/**
- * CustomerCustomerMeterSortProperty
- */
-export type CustomerCustomerMeterSortProperty =
-  | "created_at"
-  | "-created_at"
-  | "modified_at"
-  | "-modified_at"
-  | "meter_id"
-  | "-meter_id"
-  | "meter_name"
-  | "-meter_name"
-  | "consumed_units"
-  | "-consumed_units"
-  | "credited_units"
-  | "-credited_units"
-  | "balance"
-  | "-balance";
-/**
- * CustomerMeterSortProperty
- */
-export type CustomerMeterSortProperty =
-  | "created_at"
-  | "-created_at"
-  | "modified_at"
-  | "-modified_at"
-  | "customer_id"
-  | "-customer_id"
-  | "customer_name"
-  | "-customer_name"
-  | "meter_id"
-  | "-meter_id"
-  | "meter_name"
-  | "-meter_name"
-  | "consumed_units"
-  | "-consumed_units"
-  | "credited_units"
-  | "-credited_units"
-  | "balance"
-  | "-balance";
-/**
- * CustomerOrderSortProperty
- */
-export type CustomerOrderSortProperty =
-  | "created_at"
-  | "-created_at"
-  | "amount"
-  | "-amount"
-  | "net_amount"
-  | "-net_amount"
-  | "product"
-  | "-product"
-  | "subscription"
-  | "-subscription";
-/**
- * CustomerSortProperty
- */
-export type CustomerSortProperty =
-  | "created_at"
-  | "-created_at"
-  | "email"
-  | "-email"
-  | "name"
-  | "-name";
-/**
- * CustomerSubscriptionSortProperty
- */
-export type CustomerSubscriptionSortProperty =
-  | "started_at"
-  | "-started_at"
-  | "amount"
-  | "-amount"
-  | "status"
-  | "-status"
-  | "organization"
-  | "-organization"
-  | "product"
-  | "-product";
-/**
- * CustomerType
- */
-export type CustomerType = "individual" | "team";
-/**
- * CustomerWalletSortProperty
- */
-export type CustomerWalletSortProperty = "created_at" | "-created_at" | "balance" | "-balance";
-/**
- * DiscountDuration
- */
-export type DiscountDuration = "once" | "forever" | "repeating";
-/**
- * DiscountSortProperty
- */
-export type DiscountSortProperty =
-  | "created_at"
-  | "-created_at"
-  | "name"
-  | "-name"
-  | "code"
-  | "-code"
-  | "redemptions_count"
-  | "-redemptions_count"
-  | "ends_at"
-  | "-ends_at";
-/**
- * DiscountType
- */
-export type DiscountType = "fixed" | "percentage";
-/**
- * DisputeSortProperty
- */
-export type DisputeSortProperty = "created_at" | "-created_at" | "amount" | "-amount";
-/**
- * DisputeStatus
- */
-export type DisputeStatus =
-  | "prevented"
-  | "early_warning"
-  | "needs_response"
-  | "under_review"
-  | "lost"
-  | "won";
-/**
- * EventNamesSortProperty
- */
-export type EventNamesSortProperty =
-  | "name"
-  | "-name"
-  | "occurrences"
-  | "-occurrences"
-  | "first_seen"
-  | "-first_seen"
-  | "last_seen"
-  | "-last_seen";
-/**
- * EventSortProperty
- */
-export type EventSortProperty = "timestamp" | "-timestamp";
-/**
- * EventSource
- */
-export type EventSource = "system" | "user";
-/**
- * EventTypesSortProperty
- */
-export type EventTypesSortProperty =
-  | "name"
-  | "-name"
-  | "label"
-  | "-label"
-  | "occurrences"
-  | "-occurrences"
-  | "first_seen"
-  | "-first_seen"
-  | "last_seen"
-  | "-last_seen";
-/**
- * FileServiceTypes
- */
-export type FileServiceTypes =
-  | "downloadable"
-  | "product_media"
-  | "organization_avatar"
-  | "support_case_attachment";
-/**
- * FilterConjunction
- */
-export type FilterConjunction = "and" | "or";
-/**
- * FilterOperator
- */
-export type FilterOperator = "eq" | "ne" | "gt" | "gte" | "lt" | "lte" | "like" | "not_like";
-/**
- * Func
- */
-export type Func = "sum" | "max" | "min" | "avg";
-/**
- * LicenseKeyStatus
- */
-export type LicenseKeyStatus = "granted" | "revoked" | "disabled";
-/**
- * MemberRole
- */
-export type MemberRole = "owner" | "billing_manager" | "member";
-/**
- * MemberSortProperty
- */
-export type MemberSortProperty = "created_at" | "-created_at";
-/**
- * MeterSortProperty
- */
-export type MeterSortProperty = "created_at" | "-created_at" | "name" | "-name";
-/**
- * MeterUnit
- */
-export type MeterUnit = "scalar" | "token" | "custom";
-/**
- * MetricType
- */
-export type MetricType = "scalar" | "currency" | "currency_sub_cent" | "percentage";
-/**
- * OrderBillingReason
- */
-export type OrderBillingReason =
-  | "purchase"
-  | "subscription_create"
-  | "subscription_cycle"
-  | "subscription_update";
-/**
- * OrderSortProperty
- */
-export type OrderSortProperty =
-  | "created_at"
-  | "-created_at"
-  | "status"
-  | "-status"
-  | "invoice_number"
-  | "-invoice_number"
-  | "amount"
-  | "-amount"
-  | "net_amount"
-  | "-net_amount"
-  | "customer"
-  | "-customer"
-  | "product"
-  | "-product"
-  | "discount"
-  | "-discount"
-  | "subscription"
-  | "-subscription";
-/**
- * OrderStatus
- */
-export type OrderStatus = "draft" | "pending" | "paid" | "refunded" | "partially_refunded" | "void";
-/**
- * OrganizationSocialPlatforms
- */
-export type OrganizationSocialPlatforms =
-  | "x"
-  | "github"
-  | "facebook"
-  | "instagram"
-  | "youtube"
-  | "tiktok"
-  | "linkedin"
-  | "threads"
-  | "discord"
-  | "other";
-/**
- * OrganizationSortProperty
- */
-export type OrganizationSortProperty =
-  | "created_at"
-  | "-created_at"
-  | "slug"
-  | "-slug"
-  | "name"
-  | "-name"
-  | "next_review_threshold"
-  | "-next_review_threshold"
-  | "days_in_status"
-  | "-days_in_status";
-/**
- * OrganizationStatus
- */
-export type OrganizationStatus =
-  | "created"
-  | "review"
-  | "snoozed"
-  | "denied"
-  | "active"
-  | "blocked"
-  | "offboarding"
-  | "offboarded";
-/**
- * PaymentProcessor
- */
-export type PaymentProcessor = "stripe";
-/**
- * PaymentSortProperty
- */
-export type PaymentSortProperty =
-  | "created_at"
-  | "-created_at"
-  | "status"
-  | "-status"
-  | "amount"
-  | "-amount"
-  | "method"
-  | "-method";
-/**
- * PaymentStatus
- */
-export type PaymentStatus = "pending" | "succeeded" | "failed";
-/**
- * PaymentTrigger
- */
-export type PaymentTrigger =
-  | "purchase"
-  | "subscription_cycle"
-  | "retry_dunning"
-  | "retry_customer"
-  | "retry_payment_method_update"
-  | "retry_admin";
-/**
- * The permission level to grant. Read more about roles and their permissions on [GitHub documentation](https://docs.github.com/en/organizations/managing-user-access-to-your-organizations-repositories/managing-repository-roles/repository-roles-for-an-organization#permissions-for-each-role).
- */
-export type Permission = "pull" | "triage" | "push" | "maintain" | "admin";
-/**
- * PresentmentCurrency
- */
-export type PresentmentCurrency =
-  | "aed"
-  | "all"
-  | "amd"
-  | "aoa"
-  | "ars"
-  | "aud"
-  | "awg"
-  | "azn"
-  | "bam"
-  | "bbd"
-  | "bdt"
-  | "bif"
-  | "bmd"
-  | "bnd"
-  | "bob"
-  | "brl"
-  | "bsd"
-  | "bwp"
-  | "bzd"
-  | "cad"
-  | "cdf"
-  | "chf"
-  | "clp"
-  | "cny"
-  | "cop"
-  | "crc"
-  | "cve"
-  | "czk"
-  | "djf"
-  | "dkk"
-  | "dop"
-  | "dzd"
-  | "egp"
-  | "etb"
-  | "eur"
-  | "fjd"
-  | "fkp"
-  | "gbp"
-  | "gel"
-  | "gip"
-  | "gmd"
-  | "gnf"
-  | "gtq"
-  | "gyd"
-  | "hkd"
-  | "hnl"
-  | "htg"
-  | "huf"
-  | "idr"
-  | "ils"
-  | "inr"
-  | "isk"
-  | "jmd"
-  | "jpy"
-  | "kes"
-  | "kgs"
-  | "khr"
-  | "kmf"
-  | "krw"
-  | "kyd"
-  | "kzt"
-  | "lak"
-  | "lkr"
-  | "lrd"
-  | "lsl"
-  | "mad"
-  | "mdl"
-  | "mga"
-  | "mkd"
-  | "mnt"
-  | "mop"
-  | "mur"
-  | "mvr"
-  | "mwk"
-  | "mxn"
-  | "myr"
-  | "mzn"
-  | "nad"
-  | "ngn"
-  | "nio"
-  | "nok"
-  | "npr"
-  | "nzd"
-  | "pab"
-  | "pen"
-  | "pgk"
-  | "php"
-  | "pkr"
-  | "pln"
-  | "pyg"
-  | "qar"
-  | "ron"
-  | "rsd"
-  | "rwf"
-  | "sar"
-  | "sbd"
-  | "scr"
-  | "sek"
-  | "sgd"
-  | "shp"
-  | "sos"
-  | "srd"
-  | "szl"
-  | "thb"
-  | "tjs"
-  | "top"
-  | "try"
-  | "ttd"
-  | "twd"
-  | "tzs"
-  | "uah"
-  | "ugx"
-  | "usd"
-  | "uyu"
-  | "uzs"
-  | "vnd"
-  | "vuv"
-  | "wst"
-  | "xaf"
-  | "xcd"
-  | "xcg"
-  | "xof"
-  | "xpf"
-  | "yer"
-  | "zar"
-  | "zmw";
-/**
- * ProductBillingType
- */
-export type ProductBillingType = "one_time" | "recurring";
-/**
- * ProductPriceSource
- */
-export type ProductPriceSource = "catalog" | "ad_hoc";
-/**
- * ProductSortProperty
- */
-export type ProductSortProperty =
-  | "created_at"
-  | "-created_at"
-  | "name"
-  | "-name"
-  | "price_amount_type"
-  | "-price_amount_type"
-  | "price_amount"
-  | "-price_amount";
-/**
- * ProductVisibility
- */
-export type ProductVisibility = "draft" | "private" | "public";
-/**
- * PublicSubscriptionProrationBehavior
- */
-export type PublicSubscriptionProrationBehavior = "invoice" | "prorate" | "next_period";
-/**
- * Reason for the refund.
- */
-export type Reason =
-  | "duplicate"
-  | "fraudulent"
-  | "customer_request"
-  | "service_disruption"
-  | "satisfaction_guarantee"
-  | "other";
-/**
- * RecurringInterval
- */
-export type RecurringInterval = "day" | "week" | "month" | "year";
-/**
- * RefundReason
- */
-export type RefundReason =
-  | "duplicate"
-  | "fraudulent"
-  | "customer_request"
-  | "service_disruption"
-  | "satisfaction_guarantee"
-  | "dispute_prevention"
-  | "other";
-/**
- * RefundSortProperty
- */
-export type RefundSortProperty = "created_at" | "-created_at" | "amount" | "-amount";
-/**
- * RefundStatus
- */
-export type RefundStatus = "pending" | "succeeded" | "failed" | "canceled";
-/**
- * The role of the member within the customer. To assign or transfer ownership, use the member update endpoint.
- */
-export type Role = "member" | "billing_manager";
-/**
- * Scope
- */
-export type Scope =
-  | "openid"
-  | "profile"
-  | "email"
-  | "user:read"
-  | "user:write"
-  | "organizations:read"
-  | "organizations:write"
-  | "custom_fields:read"
-  | "custom_fields:write"
-  | "discounts:read"
-  | "discounts:write"
-  | "checkout_links:read"
-  | "checkout_links:write"
-  | "checkouts:read"
-  | "checkouts:write"
-  | "transactions:read"
-  | "transactions:write"
-  | "payouts:read"
-  | "payouts:write"
-  | "products:read"
-  | "products:write"
-  | "benefits:read"
-  | "benefits:write"
-  | "events:read"
-  | "events:write"
-  | "meters:read"
-  | "meters:write"
-  | "files:read"
-  | "files:write"
-  | "subscriptions:read"
-  | "subscriptions:write"
-  | "customers:read"
-  | "customers:write"
-  | "members:read"
-  | "members:write"
-  | "wallets:read"
-  | "wallets:write"
-  | "disputes:read"
-  | "disputes:write"
-  | "customer_meters:read"
-  | "customer_sessions:write"
-  | "member_sessions:write"
-  | "customer_seats:read"
-  | "customer_seats:write"
-  | "orders:read"
-  | "orders:write"
-  | "refunds:read"
-  | "refunds:write"
-  | "payments:read"
-  | "metrics:read"
-  | "metrics:write"
-  | "webhooks:read"
-  | "webhooks:write"
-  | "license_keys:read"
-  | "license_keys:write"
-  | "customer_portal:read"
-  | "customer_portal:write"
-  | "notifications:read"
-  | "notifications:write"
-  | "notification_recipients:read"
-  | "notification_recipients:write"
-  | "organization_access_tokens:read"
-  | "organization_access_tokens:write";
-/**
- * SeatStatus
- */
-export type SeatStatus = "pending" | "claimed" | "revoked";
-/**
- * SeatTierType
- */
-export type SeatTierType = "volume" | "graduated";
-/**
- * Status
- */
-export type Status = "active" | "trialing";
-/**
- * SubType
- */
-export type SubType = "user" | "organization";
-/**
- * SubscriptionProrationBehavior
- */
-export type SubscriptionProrationBehavior = "invoice" | "prorate" | "next_period" | "reset";
-/**
- * SubscriptionSortProperty
- */
-export type SubscriptionSortProperty =
-  | "customer"
-  | "-customer"
-  | "status"
-  | "-status"
-  | "started_at"
-  | "-started_at"
-  | "current_period_end"
-  | "-current_period_end"
-  | "ended_at"
-  | "-ended_at"
-  | "ends_at"
-  | "-ends_at"
-  | "amount"
-  | "-amount"
-  | "product"
-  | "-product"
-  | "discount"
-  | "-discount";
-/**
- * SubscriptionStatus
- */
-export type SubscriptionStatus =
-  | "incomplete"
-  | "incomplete_expired"
-  | "trialing"
-  | "active"
-  | "past_due"
-  | "canceled"
-  | "unpaid";
-/**
- * TaxBehavior
- */
-export type TaxBehavior = "inclusive" | "exclusive";
-/**
- * TaxBehaviorOption
- */
-export type TaxBehaviorOption = "location" | "inclusive" | "exclusive";
-/**
- * TimeInterval
- */
-export type TimeInterval = "year" | "month" | "week" | "day" | "hour";
-/**
- * Timeframe
- */
-export type Timeframe = "year" | "month" | "day";
-/**
- * Timezone to use for the timestamps. Default is UTC.
- */
-export type Timezone =
-  | "Africa/Abidjan"
-  | "Africa/Accra"
-  | "Africa/Addis_Ababa"
-  | "Africa/Algiers"
-  | "Africa/Asmara"
-  | "Africa/Asmera"
-  | "Africa/Bamako"
-  | "Africa/Bangui"
-  | "Africa/Banjul"
-  | "Africa/Bissau"
-  | "Africa/Blantyre"
-  | "Africa/Brazzaville"
-  | "Africa/Bujumbura"
-  | "Africa/Cairo"
-  | "Africa/Casablanca"
-  | "Africa/Ceuta"
-  | "Africa/Conakry"
-  | "Africa/Dakar"
-  | "Africa/Dar_es_Salaam"
-  | "Africa/Djibouti"
-  | "Africa/Douala"
-  | "Africa/El_Aaiun"
-  | "Africa/Freetown"
-  | "Africa/Gaborone"
-  | "Africa/Harare"
-  | "Africa/Johannesburg"
-  | "Africa/Juba"
-  | "Africa/Kampala"
-  | "Africa/Khartoum"
-  | "Africa/Kigali"
-  | "Africa/Kinshasa"
-  | "Africa/Lagos"
-  | "Africa/Libreville"
-  | "Africa/Lome"
-  | "Africa/Luanda"
-  | "Africa/Lubumbashi"
-  | "Africa/Lusaka"
-  | "Africa/Malabo"
-  | "Africa/Maputo"
-  | "Africa/Maseru"
-  | "Africa/Mbabane"
-  | "Africa/Mogadishu"
-  | "Africa/Monrovia"
-  | "Africa/Nairobi"
-  | "Africa/Ndjamena"
-  | "Africa/Niamey"
-  | "Africa/Nouakchott"
-  | "Africa/Ouagadougou"
-  | "Africa/Porto-Novo"
-  | "Africa/Sao_Tome"
-  | "Africa/Timbuktu"
-  | "Africa/Tripoli"
-  | "Africa/Tunis"
-  | "Africa/Windhoek"
-  | "America/Adak"
-  | "America/Anchorage"
-  | "America/Anguilla"
-  | "America/Antigua"
-  | "America/Araguaina"
-  | "America/Argentina/Buenos_Aires"
-  | "America/Argentina/Catamarca"
-  | "America/Argentina/ComodRivadavia"
-  | "America/Argentina/Cordoba"
-  | "America/Argentina/Jujuy"
-  | "America/Argentina/La_Rioja"
-  | "America/Argentina/Mendoza"
-  | "America/Argentina/Rio_Gallegos"
-  | "America/Argentina/Salta"
-  | "America/Argentina/San_Juan"
-  | "America/Argentina/San_Luis"
-  | "America/Argentina/Tucuman"
-  | "America/Argentina/Ushuaia"
-  | "America/Aruba"
-  | "America/Asuncion"
-  | "America/Atikokan"
-  | "America/Atka"
-  | "America/Bahia"
-  | "America/Bahia_Banderas"
-  | "America/Barbados"
-  | "America/Belem"
-  | "America/Belize"
-  | "America/Blanc-Sablon"
-  | "America/Boa_Vista"
-  | "America/Bogota"
-  | "America/Boise"
-  | "America/Buenos_Aires"
-  | "America/Cambridge_Bay"
-  | "America/Campo_Grande"
-  | "America/Cancun"
-  | "America/Caracas"
-  | "America/Catamarca"
-  | "America/Cayenne"
-  | "America/Cayman"
-  | "America/Chicago"
-  | "America/Chihuahua"
-  | "America/Ciudad_Juarez"
-  | "America/Coral_Harbour"
-  | "America/Cordoba"
-  | "America/Costa_Rica"
-  | "America/Coyhaique"
-  | "America/Creston"
-  | "America/Cuiaba"
-  | "America/Curacao"
-  | "America/Danmarkshavn"
-  | "America/Dawson"
-  | "America/Dawson_Creek"
-  | "America/Denver"
-  | "America/Detroit"
-  | "America/Dominica"
-  | "America/Edmonton"
-  | "America/Eirunepe"
-  | "America/El_Salvador"
-  | "America/Ensenada"
-  | "America/Fort_Nelson"
-  | "America/Fort_Wayne"
-  | "America/Fortaleza"
-  | "America/Glace_Bay"
-  | "America/Godthab"
-  | "America/Goose_Bay"
-  | "America/Grand_Turk"
-  | "America/Grenada"
-  | "America/Guadeloupe"
-  | "America/Guatemala"
-  | "America/Guayaquil"
-  | "America/Guyana"
-  | "America/Halifax"
-  | "America/Havana"
-  | "America/Hermosillo"
-  | "America/Indiana/Indianapolis"
-  | "America/Indiana/Knox"
-  | "America/Indiana/Marengo"
-  | "America/Indiana/Petersburg"
-  | "America/Indiana/Tell_City"
-  | "America/Indiana/Vevay"
-  | "America/Indiana/Vincennes"
-  | "America/Indiana/Winamac"
-  | "America/Indianapolis"
-  | "America/Inuvik"
-  | "America/Iqaluit"
-  | "America/Jamaica"
-  | "America/Jujuy"
-  | "America/Juneau"
-  | "America/Kentucky/Louisville"
-  | "America/Kentucky/Monticello"
-  | "America/Knox_IN"
-  | "America/Kralendijk"
-  | "America/La_Paz"
-  | "America/Lima"
-  | "America/Los_Angeles"
-  | "America/Louisville"
-  | "America/Lower_Princes"
-  | "America/Maceio"
-  | "America/Managua"
-  | "America/Manaus"
-  | "America/Marigot"
-  | "America/Martinique"
-  | "America/Matamoros"
-  | "America/Mazatlan"
-  | "America/Mendoza"
-  | "America/Menominee"
-  | "America/Merida"
-  | "America/Metlakatla"
-  | "America/Mexico_City"
-  | "America/Miquelon"
-  | "America/Moncton"
-  | "America/Monterrey"
-  | "America/Montevideo"
-  | "America/Montreal"
-  | "America/Montserrat"
-  | "America/Nassau"
-  | "America/New_York"
-  | "America/Nipigon"
-  | "America/Nome"
-  | "America/Noronha"
-  | "America/North_Dakota/Beulah"
-  | "America/North_Dakota/Center"
-  | "America/North_Dakota/New_Salem"
-  | "America/Nuuk"
-  | "America/Ojinaga"
-  | "America/Panama"
-  | "America/Pangnirtung"
-  | "America/Paramaribo"
-  | "America/Phoenix"
-  | "America/Port-au-Prince"
-  | "America/Port_of_Spain"
-  | "America/Porto_Acre"
-  | "America/Porto_Velho"
-  | "America/Puerto_Rico"
-  | "America/Punta_Arenas"
-  | "America/Rainy_River"
-  | "America/Rankin_Inlet"
-  | "America/Recife"
-  | "America/Regina"
-  | "America/Resolute"
-  | "America/Rio_Branco"
-  | "America/Rosario"
-  | "America/Santa_Isabel"
-  | "America/Santarem"
-  | "America/Santiago"
-  | "America/Santo_Domingo"
-  | "America/Sao_Paulo"
-  | "America/Scoresbysund"
-  | "America/Shiprock"
-  | "America/Sitka"
-  | "America/St_Barthelemy"
-  | "America/St_Johns"
-  | "America/St_Kitts"
-  | "America/St_Lucia"
-  | "America/St_Thomas"
-  | "America/St_Vincent"
-  | "America/Swift_Current"
-  | "America/Tegucigalpa"
-  | "America/Thule"
-  | "America/Thunder_Bay"
-  | "America/Tijuana"
-  | "America/Toronto"
-  | "America/Tortola"
-  | "America/Vancouver"
-  | "America/Virgin"
-  | "America/Whitehorse"
-  | "America/Winnipeg"
-  | "America/Yakutat"
-  | "America/Yellowknife"
-  | "Antarctica/Casey"
-  | "Antarctica/Davis"
-  | "Antarctica/DumontDUrville"
-  | "Antarctica/Macquarie"
-  | "Antarctica/Mawson"
-  | "Antarctica/McMurdo"
-  | "Antarctica/Palmer"
-  | "Antarctica/Rothera"
-  | "Antarctica/South_Pole"
-  | "Antarctica/Syowa"
-  | "Antarctica/Troll"
-  | "Antarctica/Vostok"
-  | "Arctic/Longyearbyen"
-  | "Asia/Aden"
-  | "Asia/Almaty"
-  | "Asia/Amman"
-  | "Asia/Anadyr"
-  | "Asia/Aqtau"
-  | "Asia/Aqtobe"
-  | "Asia/Ashgabat"
-  | "Asia/Ashkhabad"
-  | "Asia/Atyrau"
-  | "Asia/Baghdad"
-  | "Asia/Bahrain"
-  | "Asia/Baku"
-  | "Asia/Bangkok"
-  | "Asia/Barnaul"
-  | "Asia/Beirut"
-  | "Asia/Bishkek"
-  | "Asia/Brunei"
-  | "Asia/Calcutta"
-  | "Asia/Chita"
-  | "Asia/Choibalsan"
-  | "Asia/Chongqing"
-  | "Asia/Chungking"
-  | "Asia/Colombo"
-  | "Asia/Dacca"
-  | "Asia/Damascus"
-  | "Asia/Dhaka"
-  | "Asia/Dili"
-  | "Asia/Dubai"
-  | "Asia/Dushanbe"
-  | "Asia/Famagusta"
-  | "Asia/Gaza"
-  | "Asia/Harbin"
-  | "Asia/Hebron"
-  | "Asia/Ho_Chi_Minh"
-  | "Asia/Hong_Kong"
-  | "Asia/Hovd"
-  | "Asia/Irkutsk"
-  | "Asia/Istanbul"
-  | "Asia/Jakarta"
-  | "Asia/Jayapura"
-  | "Asia/Jerusalem"
-  | "Asia/Kabul"
-  | "Asia/Kamchatka"
-  | "Asia/Karachi"
-  | "Asia/Kashgar"
-  | "Asia/Kathmandu"
-  | "Asia/Katmandu"
-  | "Asia/Khandyga"
-  | "Asia/Kolkata"
-  | "Asia/Krasnoyarsk"
-  | "Asia/Kuala_Lumpur"
-  | "Asia/Kuching"
-  | "Asia/Kuwait"
-  | "Asia/Macao"
-  | "Asia/Macau"
-  | "Asia/Magadan"
-  | "Asia/Makassar"
-  | "Asia/Manila"
-  | "Asia/Muscat"
-  | "Asia/Nicosia"
-  | "Asia/Novokuznetsk"
-  | "Asia/Novosibirsk"
-  | "Asia/Omsk"
-  | "Asia/Oral"
-  | "Asia/Phnom_Penh"
-  | "Asia/Pontianak"
-  | "Asia/Pyongyang"
-  | "Asia/Qatar"
-  | "Asia/Qostanay"
-  | "Asia/Qyzylorda"
-  | "Asia/Rangoon"
-  | "Asia/Riyadh"
-  | "Asia/Saigon"
-  | "Asia/Sakhalin"
-  | "Asia/Samarkand"
-  | "Asia/Seoul"
-  | "Asia/Shanghai"
-  | "Asia/Singapore"
-  | "Asia/Srednekolymsk"
-  | "Asia/Taipei"
-  | "Asia/Tashkent"
-  | "Asia/Tbilisi"
-  | "Asia/Tehran"
-  | "Asia/Tel_Aviv"
-  | "Asia/Thimbu"
-  | "Asia/Thimphu"
-  | "Asia/Tokyo"
-  | "Asia/Tomsk"
-  | "Asia/Ujung_Pandang"
-  | "Asia/Ulaanbaatar"
-  | "Asia/Ulan_Bator"
-  | "Asia/Urumqi"
-  | "Asia/Ust-Nera"
-  | "Asia/Vientiane"
-  | "Asia/Vladivostok"
-  | "Asia/Yakutsk"
-  | "Asia/Yangon"
-  | "Asia/Yekaterinburg"
-  | "Asia/Yerevan"
-  | "Atlantic/Azores"
-  | "Atlantic/Bermuda"
-  | "Atlantic/Canary"
-  | "Atlantic/Cape_Verde"
-  | "Atlantic/Faeroe"
-  | "Atlantic/Faroe"
-  | "Atlantic/Jan_Mayen"
-  | "Atlantic/Madeira"
-  | "Atlantic/Reykjavik"
-  | "Atlantic/South_Georgia"
-  | "Atlantic/St_Helena"
-  | "Atlantic/Stanley"
-  | "Australia/ACT"
-  | "Australia/Adelaide"
-  | "Australia/Brisbane"
-  | "Australia/Broken_Hill"
-  | "Australia/Canberra"
-  | "Australia/Currie"
-  | "Australia/Darwin"
-  | "Australia/Eucla"
-  | "Australia/Hobart"
-  | "Australia/LHI"
-  | "Australia/Lindeman"
-  | "Australia/Lord_Howe"
-  | "Australia/Melbourne"
-  | "Australia/NSW"
-  | "Australia/North"
-  | "Australia/Perth"
-  | "Australia/Queensland"
-  | "Australia/South"
-  | "Australia/Sydney"
-  | "Australia/Tasmania"
-  | "Australia/Victoria"
-  | "Australia/West"
-  | "Australia/Yancowinna"
-  | "Brazil/Acre"
-  | "Brazil/DeNoronha"
-  | "Brazil/East"
-  | "Brazil/West"
-  | "CET"
-  | "CST6CDT"
-  | "Canada/Atlantic"
-  | "Canada/Central"
-  | "Canada/Eastern"
-  | "Canada/Mountain"
-  | "Canada/Newfoundland"
-  | "Canada/Pacific"
-  | "Canada/Saskatchewan"
-  | "Canada/Yukon"
-  | "Chile/Continental"
-  | "Chile/EasterIsland"
-  | "Cuba"
-  | "EET"
-  | "EST"
-  | "EST5EDT"
-  | "Egypt"
-  | "Eire"
-  | "Etc/GMT"
-  | "Etc/GMT+0"
-  | "Etc/GMT+1"
-  | "Etc/GMT+10"
-  | "Etc/GMT+11"
-  | "Etc/GMT+12"
-  | "Etc/GMT+2"
-  | "Etc/GMT+3"
-  | "Etc/GMT+4"
-  | "Etc/GMT+5"
-  | "Etc/GMT+6"
-  | "Etc/GMT+7"
-  | "Etc/GMT+8"
-  | "Etc/GMT+9"
-  | "Etc/GMT-0"
-  | "Etc/GMT-1"
-  | "Etc/GMT-10"
-  | "Etc/GMT-11"
-  | "Etc/GMT-12"
-  | "Etc/GMT-13"
-  | "Etc/GMT-14"
-  | "Etc/GMT-2"
-  | "Etc/GMT-3"
-  | "Etc/GMT-4"
-  | "Etc/GMT-5"
-  | "Etc/GMT-6"
-  | "Etc/GMT-7"
-  | "Etc/GMT-8"
-  | "Etc/GMT-9"
-  | "Etc/GMT0"
-  | "Etc/Greenwich"
-  | "Etc/UCT"
-  | "Etc/UTC"
-  | "Etc/Universal"
-  | "Etc/Zulu"
-  | "Europe/Amsterdam"
-  | "Europe/Andorra"
-  | "Europe/Astrakhan"
-  | "Europe/Athens"
-  | "Europe/Belfast"
-  | "Europe/Belgrade"
-  | "Europe/Berlin"
-  | "Europe/Bratislava"
-  | "Europe/Brussels"
-  | "Europe/Bucharest"
-  | "Europe/Budapest"
-  | "Europe/Busingen"
-  | "Europe/Chisinau"
-  | "Europe/Copenhagen"
-  | "Europe/Dublin"
-  | "Europe/Gibraltar"
-  | "Europe/Guernsey"
-  | "Europe/Helsinki"
-  | "Europe/Isle_of_Man"
-  | "Europe/Istanbul"
-  | "Europe/Jersey"
-  | "Europe/Kaliningrad"
-  | "Europe/Kiev"
-  | "Europe/Kirov"
-  | "Europe/Kyiv"
-  | "Europe/Lisbon"
-  | "Europe/Ljubljana"
-  | "Europe/London"
-  | "Europe/Luxembourg"
-  | "Europe/Madrid"
-  | "Europe/Malta"
-  | "Europe/Mariehamn"
-  | "Europe/Minsk"
-  | "Europe/Monaco"
-  | "Europe/Moscow"
-  | "Europe/Nicosia"
-  | "Europe/Oslo"
-  | "Europe/Paris"
-  | "Europe/Podgorica"
-  | "Europe/Prague"
-  | "Europe/Riga"
-  | "Europe/Rome"
-  | "Europe/Samara"
-  | "Europe/San_Marino"
-  | "Europe/Sarajevo"
-  | "Europe/Saratov"
-  | "Europe/Simferopol"
-  | "Europe/Skopje"
-  | "Europe/Sofia"
-  | "Europe/Stockholm"
-  | "Europe/Tallinn"
-  | "Europe/Tirane"
-  | "Europe/Tiraspol"
-  | "Europe/Ulyanovsk"
-  | "Europe/Uzhgorod"
-  | "Europe/Vaduz"
-  | "Europe/Vatican"
-  | "Europe/Vienna"
-  | "Europe/Vilnius"
-  | "Europe/Volgograd"
-  | "Europe/Warsaw"
-  | "Europe/Zagreb"
-  | "Europe/Zaporozhye"
-  | "Europe/Zurich"
-  | "Factory"
-  | "GB"
-  | "GB-Eire"
-  | "GMT"
-  | "GMT+0"
-  | "GMT-0"
-  | "GMT0"
-  | "Greenwich"
-  | "HST"
-  | "Hongkong"
-  | "Iceland"
-  | "Indian/Antananarivo"
-  | "Indian/Chagos"
-  | "Indian/Christmas"
-  | "Indian/Cocos"
-  | "Indian/Comoro"
-  | "Indian/Kerguelen"
-  | "Indian/Mahe"
-  | "Indian/Maldives"
-  | "Indian/Mauritius"
-  | "Indian/Mayotte"
-  | "Indian/Reunion"
-  | "Iran"
-  | "Israel"
-  | "Jamaica"
-  | "Japan"
-  | "Kwajalein"
-  | "Libya"
-  | "MET"
-  | "MST"
-  | "MST7MDT"
-  | "Mexico/BajaNorte"
-  | "Mexico/BajaSur"
-  | "Mexico/General"
-  | "NZ"
-  | "NZ-CHAT"
-  | "Navajo"
-  | "PRC"
-  | "PST8PDT"
-  | "Pacific/Apia"
-  | "Pacific/Auckland"
-  | "Pacific/Bougainville"
-  | "Pacific/Chatham"
-  | "Pacific/Chuuk"
-  | "Pacific/Easter"
-  | "Pacific/Efate"
-  | "Pacific/Enderbury"
-  | "Pacific/Fakaofo"
-  | "Pacific/Fiji"
-  | "Pacific/Funafuti"
-  | "Pacific/Galapagos"
-  | "Pacific/Gambier"
-  | "Pacific/Guadalcanal"
-  | "Pacific/Guam"
-  | "Pacific/Honolulu"
-  | "Pacific/Johnston"
-  | "Pacific/Kanton"
-  | "Pacific/Kiritimati"
-  | "Pacific/Kosrae"
-  | "Pacific/Kwajalein"
-  | "Pacific/Majuro"
-  | "Pacific/Marquesas"
-  | "Pacific/Midway"
-  | "Pacific/Nauru"
-  | "Pacific/Niue"
-  | "Pacific/Norfolk"
-  | "Pacific/Noumea"
-  | "Pacific/Pago_Pago"
-  | "Pacific/Palau"
-  | "Pacific/Pitcairn"
-  | "Pacific/Pohnpei"
-  | "Pacific/Ponape"
-  | "Pacific/Port_Moresby"
-  | "Pacific/Rarotonga"
-  | "Pacific/Saipan"
-  | "Pacific/Samoa"
-  | "Pacific/Tahiti"
-  | "Pacific/Tarawa"
-  | "Pacific/Tongatapu"
-  | "Pacific/Truk"
-  | "Pacific/Wake"
-  | "Pacific/Wallis"
-  | "Pacific/Yap"
-  | "Poland"
-  | "Portugal"
-  | "ROC"
-  | "ROK"
-  | "Singapore"
-  | "Turkey"
-  | "UCT"
-  | "US/Alaska"
-  | "US/Aleutian"
-  | "US/Arizona"
-  | "US/Central"
-  | "US/East-Indiana"
-  | "US/Eastern"
-  | "US/Hawaii"
-  | "US/Indiana-Starke"
-  | "US/Michigan"
-  | "US/Mountain"
-  | "US/Pacific"
-  | "US/Samoa"
-  | "UTC"
-  | "Universal"
-  | "W-SU"
-  | "WET"
-  | "Zulu";
-/**
- * TokenEndpointAuthMethod
- */
-export type TokenEndpointAuthMethod = "client_secret_basic" | "client_secret_post" | "none";
-/**
- * TokenType
- */
-export type TokenType = "access_token" | "refresh_token";
-/**
- * TrialInterval
- */
-export type TrialInterval = "day" | "week" | "month" | "year";
-/**
- * WebhookEventType
- */
-export type WebhookEventType =
-  | "checkout.created"
-  | "checkout.updated"
-  | "checkout.expired"
-  | "customer.created"
-  | "customer.updated"
-  | "customer.deleted"
-  | "customer.state_changed"
-  | "customer_seat.assigned"
-  | "customer_seat.claimed"
-  | "customer_seat.revoked"
-  | "member.created"
-  | "member.updated"
-  | "member.deleted"
-  | "order.created"
-  | "order.updated"
-  | "order.paid"
-  | "order.refunded"
-  | "subscription.created"
-  | "subscription.updated"
-  | "subscription.active"
-  | "subscription.canceled"
-  | "subscription.uncanceled"
-  | "subscription.revoked"
-  | "subscription.past_due"
-  | "refund.created"
-  | "refund.updated"
-  | "product.created"
-  | "product.updated"
-  | "benefit.created"
-  | "benefit.updated"
-  | "benefit_grant.created"
-  | "benefit_grant.cycled"
-  | "benefit_grant.updated"
-  | "benefit_grant.revoked"
-  | "organization.updated";
-/**
- * WebhookFormat
- */
-export type WebhookFormat = "raw" | "discord" | "slack";
