+++ sdk/typescript/src/2026-04/services/customer_portal/members.ts
@@ -1,6 +1,10 @@
 import type { ClientBase } from "../../../base";
-import type { CustomerPortalMemberCreate, CustomerPortalMemberUpdate } from "../../models/inputs";
-import type { CustomerPortalMember, ListResourceCustomerPortalMember } from "../../models/outputs";
+import type {
+  CustomerPortalMember,
+  CustomerPortalMemberCreate,
+  CustomerPortalMemberUpdate,
+  ListResourceCustomerPortalMember,
+} from "../../models";
 
 import {
   AddMember400Error,
