+++ sdk/typescript/src/2026-04/services/metrics.ts
@@ -1,7 +1,14 @@
 import type { ClientBase } from "../../base";
-import type { MetricDashboardCreate, MetricDashboardUpdate } from "../models/inputs";
-import type { ProductBillingType, TimeInterval, Timezone } from "../models/literals";
-import type { MetricDashboardSchema, MetricsLimits, MetricsResponse } from "../models/outputs";
+import type {
+  MetricDashboardCreate,
+  MetricDashboardSchema,
+  MetricDashboardUpdate,
+  MetricsLimits,
+  MetricsResponse,
+  ProductBillingType,
+  TimeInterval,
+  Timezone,
+} from "../models";
 
 import { HTTPValidationError } from "../errors";
