+++ sdk/typescript/src/2026-04/services/oauth2/clients/oauth2.ts
@@ -1,8 +1,5 @@
 import type { ClientBase } from "../../../../base";
-import type {
-  OAuth2ClientConfiguration,
-  OAuth2ClientConfigurationUpdate,
-} from "../../../models/inputs";
+import type { OAuth2ClientConfiguration, OAuth2ClientConfigurationUpdate } from "../../../models";
 
 import { HTTPValidationError } from "../../../errors";
