+++ sdk/typescript/src/2026-04/services/oauth2/index.ts
@@ -7,7 +7,7 @@ import type {
   TokenResponse,
   UserInfoOrganization,
   UserInfoUser,
-} from "../../models/outputs";
+} from "../../models";
 
 import { createClientsService } from "./clients";
